Conflicts:
	trollduction/collectors/trigger.py
	trollduction/producer.py

+++ nwcsafpps_runner/prepare_nwp.py
@@ -35,25 +35,50 @@ LOG = logging.getLogger(__name__)
 
 CONFIG_PATH = os.environ.get('PPSRUNNER_CONFIG_DIR', './')
 CONF = ConfigParser.ConfigParser()
-CONF.read(os.path.join(CONFIG_PATH, "pps_config.cfg"))
+ppsconf_path = os.path.join(CONFIG_PATH, "pps_config.cfg")
+LOG.debug("Path to config file = " + str(ppsconf_path))
+CONF.read(ppsconf_path)
 
 MODE = os.getenv("SMHI_MODE")
 if MODE is None:
     MODE = "offline"
 
+LOG.debug('MODE = ' + str(MODE))
 
 OPTIONS = {}
 for option, value in CONF.items(MODE, raw=True):
     OPTIONS[option] = value
 
 
-nhsp_path = OPTIONS.get('nhsp_path', None)
-nhsp_prefix = OPTIONS.get('nhsp_prefix', None)
+try:
+    nhsp_path = OPTIONS['nhsp_path']
+except KeyError:
+    LOG.exception('Parameter not set in config file: ' + 'nhsp_path')
+try:
+    nhsp_prefix = OPTIONS['nhsp_prefix']
+except KeyError:
+    LOG.exception('Parameter not set in config file: ' + 'nhsp_prefix')
+
 nhsf_path = OPTIONS.get('nhsf_path', None)
 nhsf_prefix = OPTIONS.get('nhsf_prefix', None)
 nwp_outdir = OPTIONS.get('nwp_outdir', None)
 nwp_lsmz_filename = OPTIONS.get('nwp_static_surface', None)
 nwp_output_prefix = OPTIONS.get('nwp_output_prefix', None)
+
+import threading
+
+
+class NwpPrepareError(Exception):
+    pass
+
+
+def logreader(stream, log_func):
+    while True:
+        s = stream.readline()
+        if not s:
+            break
+        log_func(s.strip())
+    stream.close()
 
 
 def run_command(cmdstr):
@@ -64,20 +89,21 @@ def run_command(cmdstr):
 
     LOG.debug("Command: " + str(cmdstr))
     LOG.debug('Command sequence= ' + str(myargs))
-    proc = Popen(myargs, shell=False, stderr=PIPE, stdout=PIPE)
-    while True:
-        line = proc.stdout.readline()
-        if not line:
-            break
-        LOG.info(line)
+    try:
+        proc = Popen(myargs, shell=False, stderr=PIPE, stdout=PIPE)
+    except NwpPrepareError:
+        LOG.exception("Failed when preparing NWP data for PPS...")
 
-    while True:
-        errline = proc.stderr.readline()
-        if not errline:
-            break
-        LOG.info(errline)
+    out_reader = threading.Thread(
+        target=logreader, args=(proc.stdout, LOG.info))
+    err_reader = threading.Thread(
+        target=logreader, args=(proc.stderr, LOG.info))
+    out_reader.start()
+    err_reader.start()
+    out_reader.join()
+    err_reader.join()
 
-    proc.poll()
+    return proc.returncode
 
 
 def update_nwp(starttime, nlengths):
@@ -94,7 +120,7 @@ def update_nwp(starttime, nlengths):
         LOG.info("No input files! dir = " + str(nhsf_path))
         return
 
-    LOG.debug('Files = ' + str(filelist))
+    # LOG.debug('Files = ' + str(filelist))
     for filename in filelist:
         timeinfo = filename.rsplit("_", 1)[-1]
         timestamp, step = timeinfo.split("+")
@@ -114,15 +140,32 @@ def update_nwp(starttime, nlengths):
         tmp_file = os.path.join(nwp_outdir, "tmp." + timestamp + "+" + step)
         LOG.info("result and tmp files: " +
                  str(result_file) + " " + str(tmp_file))
+        nhsp_file = os.path.join(nhsp_path, nhsp_prefix + timeinfo)
+        if not os.path.exists(nhsp_file):
+            LOG.warning("Corresponding nhsp-file not there: " + str(nhsp_file))
+            continue
+
         cmd = ("grib_copy -w gridType=regular_ll " +
-               os.path.join(nhsp_path, nhsp_prefix + timeinfo) + " " + tmp_file)
-        run_command(cmd)
+               nhsp_file + " " + tmp_file)
+        retv = run_command(cmd)
+        LOG.debug("Returncode = " + str(retv))
+
+        if not os.path.exists(nwp_lsmz_filename):
+            LOG.exception("No static grib file with land-sea mask and " +
+                          "topography available. Can't prepare NWP data")
+            raise
 
         tmpresult = tempfile.mktemp()
         cmd = ('cat ' + tmp_file + " " +
                os.path.join(nhsf_path, nhsf_prefix + timeinfo) +
                " " + nwp_lsmz_filename + " > " + tmpresult)
-        os.system(cmd)
+        LOG.debug("Add topography and land-sea mask to data:")
+        LOG.debug("Command = " + str(cmd))
+        retv = os.system(cmd)
+        LOG.debug("Returncode = " + str(retv))
+        if retv != 0:
+            raise IOError("Failed adding topography and land-sea " +
+                          "mask data to grib file")
         os.remove(tmp_file)
         os.rename(tmpresult, result_file)
